# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SomaBrain - Cognitive AI System",
    "description": "Low-latency cognitive services with strict production-mode enforcement.",
    "version": "1"
  },
  "paths": {
...
    "/persona/{pid}": {
      "put": {
        "summary": "Put Persona",
        "description": "Create or update a Persona record.\n\nUses optimistic CAS if the client supplies an If-Match header containing the\ncurrent ETag. Returns an ETag header for the newly stored representation.",
        "operationId": "put_persona_persona__pid__put",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "if-match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Persona"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Persona",
        "description": "Retrieve the latest Persona record for pid. Returns 404 if not found.",
        "operationId": "get_persona_persona__pid__get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Persona",
        "description": "Append a persona tombstone for pid. Best-effort delete for Phase 1.",
        "operationId": "delete_persona_persona__pid__delete",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Persona": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties",
            "default": {}
          },
          "fact": {
            "type": "string",
            "title": "Fact",
            "default": "persona"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "Persona",
        "description": "Persona record schema used by the Persona endpoints.\n\nAttributes:\n    id: identifier for the persona (path param)\n    display_name: optional human-readable name\n    properties: arbitrary dict with persona data\n    fact: metadata field set to 'persona' when persisted"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}